app/requirements.txt
alembic==1.11.1
sqlalchemy==2.0.17
fastapi==0.99.0
uvicorn==0.22.0
psycopg2==2.9.6
pydantic[email]==1.10.5
PyJWT==2.6.0